game storeapp
This update includes a default image icon for games that do not have a properly uploaded icon. The default image has a size of 200x200 pixels and is set using CSS for the icon element in the game list page. This ensures that there is always an image displayed for each game in the list.